==code edit 캐러셀 오류 처리 예약 내역 있으면 삭제 불가 메시지 처리

+++ healpio/src/main/java/com/healpio/service/ClassServiceImpl.java
@@ -32,9 +32,6 @@ public class ClassServiceImpl implements ClassService {
 	
 	@Autowired
 	ReviewMapper reviewMapper;
-	
-	@Autowired
-	ReservationMapper reservationMapper;
 
 	@Override
 	public void getExerciseList(Model model) {
@@ -101,6 +98,10 @@ public class ClassServiceImpl implements ClassService {
 	@Override
 	@Transactional(rollbackFor = Exception.class)
 	public int delete(String class_no, Criteria_review criteria, Model model) {
+		if(classMapper.getReservation(class_no)>0) {
+			return 0;
+		}
+		
 		// 첨부파일 삭제
 		attachService.delete(class_no);
